Merge admin branch

+++ server/database/migrations/2023_11_14_034434_create_detail_transactions_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('transaction_id')->constrained('transactions')->cascadeOnDelete();
             $table->foreignId('car_id')->constrained('cars')->cascadeOnDelete();
-            $table->integer('car_price');
+            $table->unsignedBigInteger('car_price');
             $table->integer('qty');
-            $table->integer('subtotal');
+            $table->unsignedBigInteger('subtotal');
             $table->timestamps();
         });
     }
